+++ rebuild.sh
@@ -1,0 +1,4 @@
+jlpm
+jlpm run build
+pip install . --upgrade
+jupyter lab